Refactor supplier validator handling
